refactor: Add code generation functions to test_base.py
This commit adds code generation functions to the test_base.py file. The functions generate_import_statement, get_variable_name, generate_instantiation_string, generate_call_string, and generate_script are added to facilitate code generation for testing purposes. This change improves the testability and maintainability of the codebase.
